Running test for Grow() function in HorseModel

+++ Scripts/Models/GeneDefinition.cs
@@ -6,9 +6,9 @@ public class GeneDefinition
 {
     public string GeneId {get; set;}
     public float BaseGeneStrength {get; set;}
-    public Dictionary<string, int> StatModifiers {get; set;} = new();
+    public Dictionary<string, float> StatModifiers {get; set;} = new();
 
-    public GeneDefinition(string id, float str, Dictionary<string, int> mod)
+    public GeneDefinition(string id, float str, Dictionary<string, float> mod)
     {
         GeneId = id;
         BaseGeneStrength = str;
